+++ frost-core/src/frost/round1.rs
@@ -64,7 +64,7 @@ where
     }
 
     /// Deserialize [`Nonce`] from bytes
-    pub fn from_bytes(
+    pub fn deserialize(
         bytes: <<C::Group as Group>::Field as Field>::Serialization,
     ) -> Result<Self, Error<C>> {
         <<C::Group as Group>::Field>::deserialize(&bytes)
@@ -73,7 +73,7 @@ where
     }
 
     /// Serialize [`Nonce`] to bytes
-    pub fn to_bytes(&self) -> <<C::Group as Group>::Field as Field>::Serialization {
+    pub fn serialize(&self) -> <<C::Group as Group>::Field as Field>::Serialization {
         <<C::Group as Group>::Field>::serialize(&self.0)
     }
 }
@@ -87,15 +87,6 @@ where
     }
 }
 
-// impl<C> Drop for Nonce<C>
-// where
-//     C: Ciphersuite,
-// {
-//     fn drop(&mut self) {
-//         self.zeroize()
-//     }
-// }
-
 #[cfg(any(test, feature = "test-impl"))]
 impl<C> FromHex for Nonce<C>
 where
@@ -106,7 +97,7 @@ where
     fn from_hex<T: AsRef<[u8]>>(hex: T) -> Result<Self, Self::Error> {
         let v: Vec<u8> = FromHex::from_hex(hex).map_err(|_| "invalid hex")?;
         match v.try_into() {
-            Ok(bytes) => Self::from_bytes(bytes).map_err(|_| "malformed nonce encoding"),
+            Ok(bytes) => Self::deserialize(bytes).map_err(|_| "malformed nonce encoding"),
             Err(_) => Err("malformed nonce encoding"),
         }
     }
@@ -124,14 +115,14 @@ where
     C: Ciphersuite,
 {
     /// Deserialize [`NonceCommitment`] from bytes
-    pub fn from_bytes(bytes: <C::Group as Group>::Serialization) -> Result<Self, Error<C>> {
+    pub fn deserialize(bytes: <C::Group as Group>::Serialization) -> Result<Self, Error<C>> {
         <C::Group>::deserialize(&bytes)
             .map(|element| Self(element))
             .map_err(|e| e.into())
     }
 
     /// Serialize [`NonceCommitment`] to bytes
-    pub fn to_bytes(&self) -> <C::Group as Group>::Serialization {
+    pub fn serialize(&self) -> <C::Group as Group>::Serialization {
         <C::Group>::serialize(&self.0)
     }
 }
@@ -144,7 +135,7 @@ where
     type Error = Error<C>;
 
     fn try_from(value: ElementSerialization<C>) -> Result<Self, Self::Error> {
-        Self::from_bytes(value.0)
+        Self::deserialize(value.0)
     }
 }
 
@@ -154,7 +145,7 @@ where
     C: Ciphersuite,
 {
     fn from(value: NonceCommitment<C>) -> Self {
-        Self(value.to_bytes())
+        Self(value.serialize())
     }
 }
 
@@ -164,7 +155,7 @@ where
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         f.debug_tuple("NonceCommitment")
-            .field(&hex::encode(self.to_bytes()))
+            .field(&hex::encode(self.serialize()))
             .finish()
     }
 }
@@ -197,7 +188,9 @@ where
     fn from_hex<T: AsRef<[u8]>>(hex: T) -> Result<Self, Self::Error> {
         let v: Vec<u8> = FromHex::from_hex(hex).map_err(|_| "invalid hex")?;
         match v.try_into() {
-            Ok(bytes) => Self::from_bytes(bytes).map_err(|_| "malformed nonce commitment encoding"),
+            Ok(bytes) => {
+                Self::deserialize(bytes).map_err(|_| "malformed nonce commitment encoding")
+            }
             Err(_) => Err("malformed nonce commitment encoding"),
         }
     }