+++ frost-core/src/verifying_key.rs
@@ -24,12 +24,6 @@ impl<C> VerifyingKey<C>
 where
     C: Ciphersuite,
 {
-    // pub(crate) fn from(scalar: &<<C::Group as Group>::Field as Field>::Scalar) -> Self {
-    //     let element = <C::Group as Group>::generator() * *scalar;
-
-    //     VerifyingKey { element }
-    // }
-
     /// Create a new VerifyingKey from the given element.
     #[cfg(feature = "internals")]
     pub fn new(element: <C::Group as Group>::Element) -> Self {
@@ -43,7 +37,7 @@ where
     }
 
     /// Deserialize from bytes
-    pub fn from_bytes(
+    pub fn deserialize(
         bytes: <C::Group as Group>::Serialization,
     ) -> Result<VerifyingKey<C>, Error<C>> {
         <C::Group>::deserialize(&bytes)
@@ -52,7 +46,7 @@ where
     }
 
     /// Serialize `VerifyingKey` to bytes
-    pub fn to_bytes(&self) -> <C::Group as Group>::Serialization {
+    pub fn serialize(&self) -> <C::Group as Group>::Serialization {
         <C::Group>::serialize(&self.element)
     }
 
@@ -90,7 +84,7 @@ where
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         f.debug_tuple("VerifyingKey")
-            .field(&hex::encode(self.to_bytes()))
+            .field(&hex::encode(self.serialize()))
             .finish()
     }
 }
@@ -105,7 +99,7 @@ where
     fn from_hex<T: AsRef<[u8]>>(hex: T) -> Result<Self, Self::Error> {
         let v: Vec<u8> = FromHex::from_hex(hex).map_err(|_| "invalid hex")?;
         match v.try_into() {
-            Ok(bytes) => Self::from_bytes(bytes).map_err(|_| "malformed verifying key encoding"),
+            Ok(bytes) => Self::deserialize(bytes).map_err(|_| "malformed verifying key encoding"),
             Err(_) => Err("malformed verifying key encoding"),
         }
     }
@@ -119,7 +113,7 @@ where
     type Error = Error<C>;
 
     fn try_from(value: ElementSerialization<C>) -> Result<Self, Self::Error> {
-        Self::from_bytes(value.0)
+        Self::deserialize(value.0)
     }
 }
 
@@ -129,20 +123,6 @@ where
     C: Ciphersuite,
 {
     fn from(value: VerifyingKey<C>) -> Self {
-        Self(value.to_bytes())
+        Self(value.serialize())
     }
 }
-
-// impl<C: Ciphersuite> From<VerifyingKey<C>> for <C::Group as Group>::ElementSerialization {
-//     fn from(pk: VerifyingKey<C>) -> <C::Group as Group>::ElementSerialization {
-//         pk.bytes.bytes
-//     }
-// }
-
-// impl<C: Ciphersuite> TryFrom<<C::Group as Group>::ElementSerialization> for VerifyingKey<C> {
-//     type Error = Error;
-
-//     fn try_from(bytes: [u8; 32]) -> Result<Self, Self::Error> {
-//         VerifyingKeyBytes::from(bytes).try_into()
-//     }
-// }